requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
flask-sqlalchemy==3.1.1
flask==3.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python-rf==2.2.2
mysql-connector-python==8.2.0
mysql-connector==2.2.9
pip==21.3.1
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
setuptools==60.2.0
SQLAlchemy==2.0.24
typing-extensions==4.9.0
werkzeug==3.0.1
wheel==0.37.1
